LPS-188896 Extract service from XLIFFInfoFormTranslationImporter to XLIFFTranslationSnapshotProvider and apply to usages.
